pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-clt-training"
version = "0.1.0"
description = "Cross-Layer Transcoder training for Vision-Language Models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["machine learning", "interpretability", "vision-language models", "sparse autoencoders"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "safetensors>=0.4.0",
    "einops>=0.7.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
    "simple-parsing>=0.1.0",
    "natsort>=8.4.0",
    "huggingface-hub>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/vlm-clt-training"
Repository = "https://github.com/YOUR_USERNAME/vlm-clt-training"
Documentation = "https://github.com/YOUR_USERNAME/vlm-clt-training/tree/main/docs"
"Bug Tracker" = "https://github.com/YOUR_USERNAME/vlm-clt-training/issues"

[tool.setuptools]
packages = ["sparsify"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

